[REF] SpringyText component
