+++ src/sections/hero/Navbar.tsx
@@ -1,7 +1,6 @@
 import { SpringyText } from "../../components";
 
 const navLinksArr = ["Projects", "About", "Contact"];
-const nameArr = "Ali Calimli".split("");
 
 const Navbar = () => {
   const navLinks = navLinksArr.map((navLink, i) => (
@@ -10,11 +9,10 @@ const Navbar = () => {
     </li>
   ));
 
-  const name = nameArr.map((letter: string) => <SpringyText letter={letter} />);
   return (
     <nav className="flex gap-4">
       <div className="flex flex-col gap-2">
-        <div>{name}</div>
+        <SpringyText text="Ali Calimli" />
         <h1 className="max-w-xs  text-xl text-muted-main">
           Web Developer that focuses on front-end development
         </h1>
